updated table names, removed unused modals, moved act picker
