+++ src/component/TaskTable/index.js
@@ -1,3 +1,4 @@
 export * from './TaskTable'
+export * from './ActivityTable'
 export * from './PlanTable'
-export * from './RoutineTable'
+export * from './FolderTable'
